Fix 60+ click-target bugs from full interaction audit
Sweeps the entire app after a screen-by-screen audit (see
docs/AUDIT_AND_VYAPAR_COMPARISON.md). Highlights:

CRITICAL
- Decimal cart quantities (0.5 kg loose goods) end-to-end: CartLine/
  SaleLine/BillInput/RcptLine qty are now double, with a qtyLabel formatter
  and tap-to-edit qty in the POS.
- Never sell into negative stock: addProduct/inc/setQty clamp to available;
  postSale clamps on-hand at 0; out-of-stock tiles are dimmed and refused.
  New per-item stockTracked flag (off = service) so salons still bill.
- Duplicate product name no longer overwrites the original (addStockItem
  returns null + form validation).
- Purchase duplicate-invoice guard is now enforced (button disabled).
- Receive-payment and supplier-pay clamp to the balance; exact-paise prefill;
  collect() caps at outstanding so balances can't go negative.

HIGH
- Dashboard role-gating: Create New Bill / quick actions / low-stock banner /
  Details / View-all only render when the role can reach the target (no more
  silent dead taps for cashier/accountant).
- GST dropdown snaps off-slab rates (no assertion crash on edit).
- Inventory add form validates price>0, qty/reorder>=0, unique barcode; adds
  Category/HSN/cost/service fields. Per-rate purchase GST (was flat 5%).
- purchasing _record / customers / others: capture messenger before pop.

MEDIUM/LOW
- PDF now uses bundled Inter fonts (offline reprint/share/export) and every
  fire-and-forget PDF call is wrapped with error feedback (PdfService.run).
- Disposed ~10 leaked TextEditingControllers across sheets; added mounted
  guards (scanner, lock, backup sign-out).
- Bill-discount field is controlled (clears after posting); empty-cart payment
  buttons disabled; KOT reports real print success/failure.
- Touch targets raised to >=44px: cart steppers, tap-to-edit qty, customer
  clear, template chips, Set-default, section links, trust-bar controls.
- Features Enable/Disable-all label matches unlockable caps; backup screen
  refreshes after local save/restore; tappable Recent Activity rows.
- 6 new regression tests (34 total, all passing). Bump to 0.7.0+10.

+++ billnex_app/lib/models/sale.dart
@@ -1,7 +1,7 @@
 /// A single posted line on a sale.
 class SaleLine {
   final String name;
-  final int qty;
+  final double qty; // supports weighed/loose goods (e.g. 0.5 kg)
   final double price;
   final double gstRate;
   final double discount;
@@ -11,7 +11,7 @@ class SaleLine {
   Map<String, dynamic> toJson() => {'n': name, 'q': qty, 'p': price, 'g': gstRate, 'd': discount};
   factory SaleLine.fromJson(Map<String, dynamic> j) => SaleLine(
         j['n'] as String,
-        (j['q'] as num).toInt(),
+        (j['q'] as num).toDouble(),
         (j['p'] as num).toDouble(),
         gstRate: (j['g'] as num?)?.toDouble() ?? 5,
         discount: (j['d'] as num?)?.toDouble() ?? 0,
@@ -74,7 +74,7 @@ class Sale {
     return '$hh:$mm ${d.hour < 12 ? 'AM' : 'PM'}';
   }
 
-  int get itemCount => lines.fold(0, (a, l) => a + l.qty);
+  double get itemCount => lines.fold(0.0, (a, l) => a + l.qty);
 
   Map<String, dynamic> toJson() => {
         'no': invoiceNo,
